Create FTFindUsefulData.R
This code pulls out the datsets in COSORE that contain at least 12 months of CO2 flux data with at least 15 days of data each, have 5 cm soil temperature data, and have a minimum 5 cm soil temperature below 0 oC. The name of this script was formerly FT_pulloutfunction. I changed the name to FTFindUsefulData because this was more descriptive.  @stephpenn1